Conflicts:
	src/animation/Cinematic.cpp
	src/audio/dsound/DSoundBackend.cpp
	src/core/Application.cpp
	src/core/ArxGame.cpp
	src/core/ArxGame.h
	src/core/Core.cpp
	src/game/Map.cpp
	src/game/Player.cpp
	src/graphics/opengl/OpenGLRenderer.cpp
	src/gui/Interface.cpp
	src/gui/Menu.cpp
	src/input/DInput8Backend.cpp
	src/input/Input.cpp
	src/scene/ChangeLevel.cpp
	src/scene/LoadLevel.cpp
	src/window/D3D9Window.cpp
	src/window/RenderWindow.cpp
	src/window/RenderWindow.h
	src/window/SDLWindow.cpp
	src/window/SDLWindow.h
	src/window/Win32Window.cpp
	src/window/Win32Window.h
	src/window/Window.cpp
	src/window/Window.h

+++ tools/crashreporter/tbg/TBG.cpp
@@ -198,7 +198,7 @@ bool Server::findIssue(const QString& text, int& issue_id)
 		// add a dependency to QtXmlPatterns...
 		// Look for the first "/rss/channel/item/link" entry...
 		const QString XML_PATH[] = { "rss", "channel", "item", "link" };
-		int currentItem = 0;
+		size_t currentItem = 0;
 		QString issueLink;
 
 		while(!xml.atEnd() && !xml.hasError())
@@ -206,7 +206,7 @@ bool Server::findIssue(const QString& text, int& issue_id)
 			// Read next element
 			QXmlStreamReader::TokenType token = xml.readNext();
 
-			if(currentItem == boost::size(XML_PATH))
+			if(currentItem == size_t(boost::size(XML_PATH)))
 			{
 				issueLink = xml.text().toString();
 				break;
